+++ components/modal/addModal/AddLapModal.tsx
@@ -1,5 +1,5 @@
 import Button from '@/components/BaseButton';
-import CustomDateTimePicker from '@/components/CustomDatetimePicker';
+import CustomDateTimePicker from '@/components/modal/CustomDatetimePicker';
 import { AddableLap, AddableLapModalProp } from '@/src/types/AddableTravels';
 import React, { useState } from 'react';
 import {
